third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-absolute-child-with-percent-left-inside-relative-parent-expected.txt

Blue item box is 10% left in container. Box should display 40px as getComputedStyle( item ).left should return 40px (400px * 10%)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('left') is '40px'
PASS successfullyParsed is true

TEST COMPLETE
